move progresspoll to its own file
